Separate path watcher and scheduler
