+++ airsonic-main/src/main/java/org/airsonic/player/service/TaskSchedulingService.java
@@ -23,16 +23,7 @@ import org.springframework.scheduling.support.PeriodicTrigger;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 
-import javax.annotation.PostConstruct;
-
 import java.io.IOException;
-import java.nio.file.FileSystems;
-import java.nio.file.Path;
-import java.nio.file.StandardWatchEventKinds;
-import java.nio.file.WatchEvent;
-import java.nio.file.WatchEvent.Kind;
-import java.nio.file.WatchKey;
-import java.nio.file.WatchService;
 import java.time.Duration;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
@@ -41,11 +32,8 @@ import java.util.Date;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
-import java.util.Optional;
 import java.util.Set;
-import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.function.BiConsumer;
 import java.util.function.Function;
 
 import static java.util.stream.Collectors.toMap;
@@ -53,37 +41,12 @@ import static java.util.stream.Collectors.toMap;
 @Service
 public class TaskSchedulingService implements ScheduledTaskHolder {
     private static final Logger LOG = LoggerFactory.getLogger(TaskSchedulingService.class);
-    private ScheduledTaskRegistrar registrar;
+
+    private final ScheduledTaskRegistrar registrar;
 
     private final Map<String, ScheduledTask> tasks = new ConcurrentHashMap<>();
     private final Map<ScheduledTask, Map<String, Object>> taskMetadata = new ConcurrentHashMap<>();
-    private final Map<WatchKey, Map<Kind<? extends Object>, BiConsumer<Path, WatchEvent<Path>>>> watchFunctions = new ConcurrentHashMap<>();
-    private final Map<String, WatchKey> watchNames = new ConcurrentHashMap<>();
-    private WatchService watchService;
-    private final Runnable watcherTask = () -> {
-        for (;;) {
-            WatchKey key;
-            try {
-                key = watchService.take();
-            } catch (Exception e) {
-                LOG.warn("Watch service unable to poll. Directory monitoring may be off", e);
-                continue;
-            }
-            Map<Kind<? extends Object>, BiConsumer<Path, WatchEvent<Path>>> fs = watchFunctions.get(key);
-            if (fs == null) {
-                continue;
-            }
-            for (WatchEvent<?> event : key.pollEvents()) {
-                Optional.ofNullable(fs.get(event.kind()))
-                        .ifPresent(f -> scheduleOnce("path-watcher-function-execution",
-                                () -> f.accept((Path) key.watchable(), (WatchEvent<Path>) event),
-                                Instant.now(),
-                                false));
-            }
-            // To keep receiving events
-            key.reset();
-        }
-    };
+
 
     @Autowired
     public TaskSchedulingService(TaskSchedulerBuilder builder) throws IOException {
@@ -91,15 +54,7 @@ public class TaskSchedulingService implements ScheduledTaskHolder {
         taskScheduler.setDaemon(true);
         taskScheduler.afterPropertiesSet();
         this.registrar = new ScheduledTaskRegistrar();
-        registrar.setScheduler(taskScheduler);
-        this.watchService = FileSystems.getDefault().newWatchService();
-    }
-
-    @PostConstruct
-    public void init() {
-        scheduleOnce("path-watcher", watcherTask, Instant.now(), true);
-        // this.registrar.scheduleTriggerTask(new TriggerTask(watcherTask, new
-        // RunOnceTrigger(0L)));
+        this.registrar.setScheduler(taskScheduler);
     }
 
     public void scheduleTask(String name, Function<ScheduledTaskRegistrar, ScheduledTask> scheduledTask) {
@@ -143,56 +98,6 @@ public class TaskSchedulingService implements ScheduledTaskHolder {
         ScheduledTask task = tasks.remove(name);
         if (task != null) {
             task.cancel();
-        }
-    }
-
-    public void setWatcher(Path watchable, BiConsumer<Path, WatchEvent<Path>> entryCreateConsumer,
-            BiConsumer<Path, WatchEvent<Path>> entryDeleteConsumer,
-            BiConsumer<Path, WatchEvent<Path>> entryModifyConsumer, BiConsumer<Path, WatchEvent<Path>> overflowConsumer)
-            throws IOException {
-        setWatcher(UUID.randomUUID().toString(), watchable, entryCreateConsumer, entryDeleteConsumer, entryModifyConsumer, overflowConsumer);
-    }
-
-    public void setWatcher(String id, Path watchable, BiConsumer<Path, WatchEvent<Path>> entryCreateConsumer,
-            BiConsumer<Path, WatchEvent<Path>> entryDeleteConsumer,
-            BiConsumer<Path, WatchEvent<Path>> entryModifyConsumer, BiConsumer<Path, WatchEvent<Path>> overflowConsumer)
-            throws IOException {
-        Map<Kind<? extends Object>, BiConsumer<Path, WatchEvent<Path>>> fnMap = new HashMap<>();
-        fnMap.put(StandardWatchEventKinds.ENTRY_CREATE, entryCreateConsumer);
-        fnMap.put(StandardWatchEventKinds.ENTRY_DELETE, entryDeleteConsumer);
-        fnMap.put(StandardWatchEventKinds.ENTRY_MODIFY, entryModifyConsumer);
-        fnMap.put(StandardWatchEventKinds.OVERFLOW, overflowConsumer);
-        setWatcher(id, watchable, fnMap);
-    }
-
-    public void setWatcher(String id, Path watchable,
-            Map<Kind<? extends Object>, BiConsumer<Path, WatchEvent<Path>>> fnMap) throws IOException {
-        WatchKey key = watchable.register(watchService, StandardWatchEventKinds.ENTRY_CREATE, StandardWatchEventKinds.ENTRY_DELETE, StandardWatchEventKinds.ENTRY_MODIFY, StandardWatchEventKinds.OVERFLOW);
-        watchFunctions.put(key, fnMap);
-        watchNames.compute(id, (k, v) -> {
-            if (!key.equals(v)) {
-                invalidateWatcher(v);
-            }
-            return key;
-        });
-    }
-
-    public void invalidateWatcher(String id) {
-        WatchKey key = watchNames.remove(id);
-        invalidateWatcher(key);
-    }
-
-    public void invalidateWatcher(Path watchable) {
-        watchNames.entrySet().stream()
-                .filter(e -> e.getValue().watchable().equals(watchable))
-                .findAny()
-                .ifPresent(e -> invalidateWatcher(e.getKey()));
-    }
-
-    public void invalidateWatcher(WatchKey key) {
-        if (key != null) {
-            key.cancel();
-            watchFunctions.remove(key);
         }
     }
